Add files via upload
This MATLAB code can be used to import the velocity time series created using the OpenFOAM (post)Processing utility "probe". The location of the probes are indicated in a file, e.g. "probes" placed in the system folder. The function must also be invoked in the controlDict file, also in the system folder. Once the solver starts, a new folder is created called "postProcessing", which will in turn contain folders of the (post)Processing utilities invoked in controlDict. Inside the "probes" folder will be one or more folders, each of which is the starting time of the runs. If run only once, starting from "0", then there will only be a "0" folder. Inside each of these time folders will be a file called "p" (pressure) and "U" (velocity) and/or whatever else is indicated in the "probes" file in the system folder. This MATLAB script can be used to import the velocity time series data from "U", and it could easily be modified to import from "p" as well.