Add files via upload
This MATLAB script can be used to import the entire pressure and velocity field from a single time from the result of an OpenFOAM simulation. Before running the script, one must run the "writeCellCentres" utility to determine the locations of the cell centres, and indicate in the script where these files are. See the script for more information. One can also choose whether to interpolate or not.